RestartButton and SpriteAtlas to fix tearing
A simple RestartButton script and asset was added to the TestScene (inside a UI Canvas object) to test out resetting the scene.

Issues were found with the PlayerMovementController's inputs' events not unsubscribing when scene is changed, so actual callbacks are used like in PlayerPowerController.

A Sprite Atlas was created to contain the sprites for the used tilesets in order to fix the issue with visual tearing when moving across the environment.