Remade PlayerInputController + added RebindActionUI prefab
Remade the PlayerInputController script to now depend on a directly given InputActionAsset scriptable object instead of the C# generated class for the input controls (PlayerControls in this case).

This is because the PlayerControls C# generated class is disconnected from the PlayerControls InputActionAsset (as it keeps a copy of the JSON to generate it's own asset internally), and thus is unaffected by the regular way of rebinding action controls.

A modified version of Unity's RebindUI sample's prefab was made (now using TextMeshPro), and tested successfully. The prefab is named RebindActionUI.