Created Prefab Palettes + LevelTemplate scene + PowerBarUI prefab modified WaterGrowablePlant bugfix + General folder and file ordering
Created Interactable, Items, Entities, and MovableObjects prefab palettes to make level design easier.

An empty LevelTemplate (with all the necessary GameObjects) was created.

Bugfix for the WaterGrowableObject not disappearing after spawning the vine/ladder.

Made the PowerBarUI prefab a panel with children objects, instead of a whole canvas. This allows easily adding it to the GameUI canvas, instead of existing on it's own.

Some general ordering of folders and files, as well as prefab polishing (putting Prefab default coordinates back to 0 instead of thousands of units away for some).