LevelManager optimisation + Increased Size of PowerBarUI
Extracted the code that used FindObjectsOfType(...) to populate treeArray and ghostArray in LevelManager's Update() function. That function is costly, and shouldn't be called every frame. A single call to populate it in Start() suffices.

Increased the size of the PowerBarUI, to make it easier for players to see how many charges they have with a quick look.